Move FileSinkConfig out of Core
